+++ tests/unit/deploy/test_deployer.py
@@ -1228,6 +1228,7 @@ class TestSwaggerBuilder(object):
             resource_name='foo',
             swagger_doc=models.Placeholder.BUILD_STAGE,
             minimum_compression='',
+            endpoint_type='EDGE',
             api_gateway_stage='api',
             lambda_function=None,
         )