+++ tests/unit/deploy/test_validate.py
@@ -12,6 +12,7 @@ from chalice.deploy.validate import validate_python_version
 from chalice.deploy.validate import validate_route_content_types
 from chalice.deploy.validate import validate_unique_function_names
 from chalice.deploy.validate import validate_feature_flags
+from chalice.deploy.validate import validate_endpoint_type
 from chalice.deploy.validate import ExperimentalFeatureError
 
 
@@ -231,6 +232,21 @@ def test_can_validate_updated_custom_binary_types(sample_app):
 
     assert validate_route_content_types(sample_app.routes,
                                         sample_app.api.binary_types) is None
+
+
+def test_can_validate_endpoint_type(sample_app):
+    config = Config.create(
+        chalice_app=sample_app, api_gateway_endpoint_type='EDGE2')
+    with pytest.raises(ValueError):
+        validate_endpoint_type(config)
+
+    config = Config.create(
+        chalice_app=sample_app, api_gateway_endpoint_type='REGIONAL')
+    validate_endpoint_type(config)
+
+    config = Config.create(
+        chalice_app=sample_app, api_gateway_endpoint_type='PRIVATE')
+    validate_endpoint_type(config)
 
 
 def test_can_validate_feature_flags(sample_app):