+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Next Release (TBD)
 * Add experimental support for websockets
   (`#1017 <https://github.com/aws/chalice/issues/1017>`__)
 
+* API Gateway Endpoint Type Configuration (Regional and Edge)
+  (`#1160 https://github.com/aws/chalice/pull/1160`__)
+
 
 1.9.1
 =====